+++ NEWS.md
@@ -1,6 +1,7 @@
 # spant 2.18.0
 * Improved automated file type detection.
 * Bug fixes for DICOM MRS reader.
+* scale_spec function now works with basis set objects.
 
 # spant 2.17.0
 * Simulation duration output is now rounded to 2dp.